+++ resources/views/app.blade.php
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0" />
-    <title>SWIRL</title>
+    <title>Swirl</title>
     @vite('resources/js/app.tsx')
     @inertiaHead
     <link rel="preload" href="https://rsms.me/inter/inter.css" as="style" />