+++ vite.config.js
@@ -1,3 +1,4 @@
+import tailwindcss from '@tailwindcss/vite'
 import laravel from 'laravel-vite-plugin'
 import { defineConfig } from 'vite'
 
@@ -6,6 +7,7 @@ export default defineConfig({
     laravel({
       input: ['resources/css/app.css', 'resources/js/app.tsx'],
       refresh: true
-    })
+    }),
+    tailwindcss()
   ]
 })